maybe basic done

+++ spx_exchange.c
@@ -7,6 +7,7 @@
 #include "spx_exchange.h"
 #include <fcntl.h>
 #include <signal.h>
+#include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -24,11 +25,17 @@ static inline int min(int a, int b)
 
 trader_node_t *head = NULL;
 trader_node_t *tail = NULL;
-order_node_t *order_head = NULL;
-order_node_t *order_tail = NULL;
+order_node_t *buy_head = NULL;
+order_node_t *buy_tail = NULL;
+order_node_t *sell_head = NULL;
+order_node_t *sell_tail = NULL;
+order_book_t *book = NULL;
+order_book_t *book_tail = NULL;
 
 int trader_count = 0;
 int fees = 0;
+
+trader_node_t *current;
 
 #define MAX_EVENTS 10
 int epfd = -1;
@@ -47,13 +54,21 @@ void signal_handler(int sig)
     for (int i = 0; i < nfds; ++i) {
         // look up events[n].data.fd and send
         int fd = events[i].data.fd;
-        trader_node_t *node = trader_lookup_fd(fd);
-        assert(node != NULL);
+        current = trader_lookup_fd(fd);
+        assert(current != NULL);
         int bytes = read(fd, buffer, 128);
         if (bytes == 0) {
-            print_teardown(node->id);
+            // pipe error
+            print_teardown(current->id);
+            assert (epoll_ctl(epfd, EPOLL_CTL_DEL, current->trader_fd, &ev) != -1);
+            assert (epoll_ctl(epfd, EPOLL_CTL_DEL, current->exchagne_fd, &ev) != -1);
+            close(current->exchagne_fd);
+            close(current->trader_fd);
+            current->valid = 0;
+            continue;
         }
-        
+
+        spx_log("[T%d] Parsing command: <%s>\n", current->id, buffer);
         // split the message
         order_node_t new_order;
         char *token = strtok(buffer, " ");
@@ -68,15 +83,22 @@ void signal_handler(int sig)
         new_order.price = atoi(token);
 
         // process
-        new_order.trader_id = node->id;
+        new_order.trader_id = current->id;
 
-        order_node_t *order_check = order_lookup(new_order.order_id);
+        order_node_t *order_check = NULL;
+        if (strstr(new_order.type, "BUY"))
+            order_check = order_lookup(buy_head, new_order.order_id);
+        else
+            order_check = order_lookup(sell_head, new_order.order_id);
         if (order_check) {
             // cancelled if num and price == 0
             // amended if order_id is the same
             if (new_order.num == 0 && new_order.price == 0) {
                 // delete the node
-                order_remove(new_order.order_id);
+                if (strstr(new_order.type, "BUY"))
+                    order_remove(&buy_head, &buy_tail, new_order.order_id);
+                else
+                    order_remove(&sell_head, &sell_tail, new_order.order_id);
                 sprintf(message, "CANCELLED %d;", new_order.order_id);
             } else {
                 order_check->num = new_order.num;
@@ -88,21 +110,34 @@ void signal_handler(int sig)
             sprintf(message, "ACCEPTED %d;", new_order.order_id);
         }
 
-        write(node->exchagne_fd, message, 128);
-        kill(node->pid, SIGUSR1);
+        write(current->exchagne_fd, message, 128);
+        kill(current->pid, SIGUSR1);
 
         // notify_except
         sprintf(message, "MARKET %s %s %d %d;", new_order.type, new_order.product, new_order.num, new_order.price);
-        notify_except(node->id, buffer);
+        notify_except(current->id, buffer);
 
         // match order_list
         match_orders();
+        report();
     }
+}
+
+void pipe_handler(int sig)
+{
+    print_teardown(current->id);
+    assert (epoll_ctl(epfd, EPOLL_CTL_DEL, current->trader_fd, &ev) != -1);
+    assert (epoll_ctl(epfd, EPOLL_CTL_DEL, current->exchagne_fd, &ev) != -1);
+    close(current->exchagne_fd);
+    close(current->trader_fd);
+    current->valid = 0;
 }
 
 int main(int argc, char **argv) {
     epfd = epoll_create1(0);
     signal(SIGUSR1, signal_handler);
+    signal(SIGPIPE, pipe_handler);
+
     startup(argc, argv);
     wait_all();
     clean_all();
@@ -117,14 +152,27 @@ void startup(int argc, char **argv)
                 "./spx_exchange [product file] [trader 0] [trader 1] ... [trader n]\n");
         exit(1);
     }
+
     FILE* products = fopen(argv[1], "r");
     if (products == NULL) {
         fprintf(stderr, "file %s cannot be open\n", argv[1]);
         exit(2);
     }
+    spx_log("Starting\n");
+
     char buffer[128];
     fgets(buffer, 128, products);
     trader_count = atoi(buffer);
+    spx_log("Trading %d products: ", trader_count);
+    for (int i = 0; i < trader_count; ++i) {
+        fgets(buffer, 128, products);
+        buffer[strlen(buffer) - 1] = 0; // remove endl
+        add_products(buffer);
+        if (i != trader_count - 1)
+            printf("%s ", buffer);
+        else
+            printf("%s\n", buffer);
+    }
     fclose(products);
 
     // name pipes
@@ -133,12 +181,13 @@ void startup(int argc, char **argv)
         char name[32];
         sprintf(name, FIFO_EXCHANGE, i);
         mkfifo(name, 0666);
+        spx_log("Created FIFO %s\n", name);
 
         sprintf(name, FIFO_TRADER, i);
         mkfifo(name, 0666);
+        spx_log("Created FIFO %s\n", name);
     }
 
-    // child process
     for (int i = 2; i < argc; ++i) {
         pid_t pid = fork();
         if (pid == 0) {
@@ -150,8 +199,9 @@ void startup(int argc, char **argv)
             assert(0);
         }
 
-        // father
+        // parent
         // add pid
+        spx_log("Starting trader %d (%s)\n", i - 2, argv[i]);
         add_trader_node(i - 2, pid);
     }
 
@@ -159,17 +209,60 @@ void startup(int argc, char **argv)
     notify_all(message);
 }
 
+void add_products(char *name)
+{
+    order_book_t *new_product = calloc(1, sizeof(order_book_t));
+    strcpy(new_product->name, name);
+
+    if (book == NULL) {
+        book_tail = book = new_product;
+    } else {
+        book_tail->next = new_product;
+        book_tail = new_product;
+    }
+}
+
+void clean_book(order_book_t *head)
+{
+    order_book_t *node = head;
+    while (node) {
+        order_book_t *backup = node;
+        node = node->next;
+        free(backup);
+    }
+}
+
+order_book_t* copy_book()
+{
+    order_book_t *old = book, *new = calloc(1, sizeof(order_book_t));
+    order_book_t *tail = new;
+    while (old) {
+        memcpy(tail, old, sizeof(order_book_t));
+        old = old->next;
+        if (old) {
+            tail->next = calloc(1, sizeof(order_book_t));
+            tail = tail->next;
+        }
+    }
+    return new;
+}
+
 void add_trader_node(int id, pid_t pid)
 {
     trader_node_t *new_node = calloc(1, sizeof(trader_node_t));
     new_node->id = id;
     new_node->pid = pid;
+    new_node->valid = 1;
+    new_node->book = copy_book();
+
     char name[32];
     sprintf(name, FIFO_EXCHANGE, id);
     new_node->exchagne_fd = open(name, O_WRONLY);
+    spx_log("Connected to %s\n", name);
 
     sprintf(name, FIFO_TRADER, id);
     new_node->trader_fd = open(name, O_RDONLY);
+    spx_log("Connected to %s\n", name);
 
     // add to signal handler
     struct epoll_event ev;
@@ -214,8 +307,10 @@ void notify_all(char *message)
 {
     trader_node_t *node = head;
     while (node) {
-        write(node->exchagne_fd, message, 128);
-        kill(node->pid, SIGUSR1);
+        if (node->valid) {
+            write(node->exchagne_fd, message, 128);
+            kill(node->pid, SIGUSR1);
+        }
         node = node->next;
     }
 }
@@ -224,8 +319,10 @@ void notify_except(int id, char *message)
 {
     trader_node_t *node = head;
     while (node) {
-        if (node->id == id)
+        if (node->id == id || !node->valid) {
+            node = node->next;
             continue;
+        }
         write(node->exchagne_fd, message, 128);
         kill(node->pid, SIGUSR1);
         node = node->next;
@@ -235,9 +332,8 @@ void notify_except(int id, char *message)
 void wait_all()
 {
     trader_node_t *node = head;
-    int status;
     while (node) {
-        waitpid(node->pid, &status, 0);
+        waitpid(node->pid, NULL, 0);
         node = node->next;
     }
 }
@@ -248,24 +344,46 @@ void clean_all()
     while (node) {
         trader_node_t *backup = node;
         node = node->next;
+        if (backup->valid) {
+            close(backup->exchagne_fd);
+            close(backup->trader_fd);
+        }
+        clean_book(backup->book);
         free(backup);
     }
+    clean_book(book);
+    clean_orders(sell_head);
+    clean_orders(buy_head);
+
+    spx_log("Trading completed\n");
+    spx_log("Exchange fees collected: $%d\n", fees);
+}
+
+
+void add_order_node(order_node_t **head, order_node_t **tail, order_node_t *new_node)
+{
+    if (*tail == NULL) {
+        *head = *tail = new_node;
+        return;
+    }
+
+    (*tail)->next = new_node;
+    new_node->prev = *tail;
+    *tail = new_node;
 }
 
 void add_order(order_node_t *node)
 {
     order_node_t *new_node = calloc(1, sizeof(order_node_t));
     memcpy(new_node, node, sizeof(order_node_t));
-    if (order_tail == NULL) {
-        order_head = order_tail = new_node;
-        return;
+    if (strstr(new_node->type, "BUY")) {
+        add_order_node(&buy_head, &buy_tail, new_node);
+    } else {
+        add_order_node(&sell_head, &sell_tail, new_node);
     }
-
-    order_tail->next = new_node;
-    order_tail = new_node;
 }
 
-order_node_t *order_lookup(int id)
+order_node_t *order_lookup(order_node_t *order_head, int id)
 {
     order_node_t *node = order_head;
     while (node) {
@@ -277,25 +395,25 @@ order_node_t *order_lookup(int id)
     return NULL;
 }
 
-void order_remove(int id)
+void order_remove(order_node_t **order_head, order_node_t **order_tail, int id)
 {
-    if (order_head->order_id == id) {
-        order_node_t *backup = order_head;
-        order_head = order_head->next;
-        if (order_head == NULL) {
-            order_tail = NULL;
+    if ((*order_head)->order_id == id) {
+        order_node_t *backup = *order_head;
+        *order_head = (*order_head)->next;
+        if (*order_head == NULL) {
+            *order_tail = NULL;
         }
         free(backup);
         return;
     }
 
-    order_node_t *node = order_head;
+    order_node_t *node = *order_head;
     while (node->next) {
         if (node->next->order_id == id) {
             order_node_t *backup = node->next;
             node->next = backup->next;
-            if (backup == order_tail) {
-                order_tail = NULL;
+            if (backup == *order_tail) {
+                *order_tail = NULL;
             }
             free(backup);
             return;
@@ -304,59 +422,241 @@ void order_remove(int id)
     }
 }
 
-// a is the first and b is the last
+// return true if b is empty
 bool deal_order(order_node_t *a, order_node_t *b, int money)
 {
     int deal_num = min(a->num, b->num);
     a->num -= deal_num;
     b->num -= deal_num;
-    
+
     // fee
     int total = money * deal_num;
-    int fee = (total + 99) / 100;
+    int fee = (total + 50) / 100;
     fees += fee;
 
     // fill
     char fill_message[128];
     sprintf(fill_message, "FILL %d %d;", a->trader_id, deal_num);
     trader_node_t *trader = trader_lookup_id(a->trader_id);
+    // fees
+    order_book_t *trader_book = trader->book;
+    while (trader_book) {
+        if (strcmp(trader_book->name, a->product) == 0) {
+            if (strstr(a->type, "BUY")) {
+                trader_book->num += deal_num;
+                trader_book->total_price -= total;
+            }
+            else {
+                trader_book->num -= deal_num;
+                trader_book->total_price += total;
+            }
+            break;
+        }
+        trader_book = trader_book->next;
+    }
+
     write(trader->exchagne_fd, fill_message, 128);
     kill(trader->pid, SIGUSR1);
 
     sprintf(fill_message, "FILL %d %d;", b->trader_id, deal_num);
     trader = trader_lookup_id(b->trader_id);
+    // fees
+    trader_book = trader->book;
+    while (trader_book) {
+        if (strcmp(trader_book->name, b->product) == 0) {
+            if (strstr(b->type, "BUY")) {
+                trader_book->num += deal_num;
+                trader_book->total_price -= total + fee;
+            }
+            else {
+                trader_book->num -= deal_num;
+                trader_book->total_price += total - fee;
+            }
+            break;
+        }
+        trader_book = trader_book->next;
+    }
     write(trader->exchagne_fd, fill_message, 128);
     kill(trader->pid, SIGUSR1);
+    
+    spx_log("Match: Order %d [T%d], New Order %d [T%d], value: $%d, fee: $%d.\n", 
+            a->order_id, a->trader_id, b->order_id, b->trader_id, total, fee);
 
     if (a->num == 0) {
-        order_remove(a->order_id);
+        if (strstr(a->type, "BUY"))
+            order_remove(&buy_head, &buy_head, a->order_id);
+        else
+            order_remove(&sell_head, &sell_head, a->order_id);
     }
 
     if (b->num == 0) {
-        order_remove(b->order_id);
+        if (strstr(a->type, "BUY"))
+            order_remove(&buy_head, &buy_head, b->order_id);
+        else
+            order_remove(&sell_head, &sell_head, b->order_id);
         return true;
     }
-    
+
     return false;
+}
+
+order_node_t* copy_and_sort(order_node_t *head)
+{
+    order_node_t *old = head, *new = calloc(1, sizeof(order_node_t));
+    order_node_t *tail = new;
+
+    while (old) {
+        memcpy(tail, old, sizeof(order_node_t));
+        old = old->next;
+        if (old) {
+            tail->next = calloc(1, sizeof(order_node_t));
+            tail = tail->next;
+        }
+    }
+
+    // sort new head
+    order_node_t* tempval;
+    order_node_t *node = new->next, *key = NULL;
+    // 插入排序
+    while (node) {
+        key = node;
+        while (key != NULL && key->prev != NULL && key->price > key->prev->price) {
+            // swap key and key->prev
+            tempval = key->prev;
+            key->prev = tempval->prev;
+            tempval->next = key->next;
+
+            if (tempval->prev)
+                tempval->prev->next = key;
+            tempval->prev = key;
+
+            if (key->next)
+                key->next->prev = tempval;
+            key->next = tempval;
+            key = key->prev;
+        }
+        node = node->next;
+    }
+
+    return new;
+}
+
+void clean_orders(order_node_t *head)
+{
+    order_node_t *node = head;
+    while (node) {
+        order_node_t *backup = node;
+        node = node->next;
+        free(backup);
+    }
+}
+
+void report()
+{
+    spx_log("\t--ORDERBOOK--\n");
+    // traverse orderbook
+    order_book_t *node = book;
+
+    // sort
+    if (buy_tail)
+        buy_tail->next = sell_head;
+    if (sell_head)
+        sell_head->prev = buy_tail;
+
+    order_node_t *buy_and_sells = copy_and_sort(buy_head);
+    while (node) {
+        spx_log("\tProduct: %s; ", node->name);
+        int buy_levels = 0, sell_levels = 0;
+
+        // get levels
+        order_node_t *temp = buy_and_sells;
+        while (temp) {
+            if (strcmp(temp->product, node->name))
+            {
+                temp = temp->next;
+                continue;
+            }
+            if (temp->next && temp->next->price == temp->price && strcmp(temp->product, temp->next->product) == 0) {
+                temp = temp->next;
+                continue;
+            }
+
+            if (strstr(temp->type, "BUY"))
+                buy_levels++;
+            else
+                sell_levels++;
+            temp = temp->next;
+        }
+        printf("Buy levels: %d; Sell levels: %d\n", buy_levels, sell_levels);
+
+        temp = buy_and_sells;
+        int orders = 0;
+        int nums = 0;
+        while (temp) {
+            if (strcmp(temp->product, node->name))
+            {
+                temp = temp->next;
+                continue;
+            }
+
+            if (temp->next && temp->next->price == temp->price && strcmp(temp->product, temp->next->product) == 0) {
+                temp = temp->next;
+                orders++;
+                nums += temp->num;
+                continue;
+            }
+
+            orders++;
+            nums += temp->num;
+            if (orders > 1)
+                spx_log("\t\t%s %d @ $%d (%d orders)\n", temp->type, nums, temp->price, orders);
+            else
+                spx_log("\t\t%s %d @ $%d (%d order)\n", temp->type, nums, temp->price, orders);
+
+            temp = temp->next;
+        }
+        node = node->next;
+    }
+    clean_orders(buy_and_sells);
+    if (buy_tail)
+        buy_tail->next = NULL;
+    if (sell_head)
+        sell_head->prev = NULL;
+
+    spx_log("\t--POSITIONS--\n");
+    trader_node_t *trader = head;
+    while (trader) {
+        spx_log("\t\tTrader %d: ", trader->id);
+        order_book_t *trader_book = trader->book;
+        while (trader_book) {
+            if (trader_book->next)
+                printf("%s %d ($%d), ", trader_book->name, trader_book->num, trader_book->total_price);
+            else
+                printf("%s %d ($%d)\n", trader_book->name, trader_book->num, trader_book->total_price);
+            trader_book = trader_book->next;
+        }
+        trader = trader->next;
+    }
 }
 
 void match_orders()
 {
-    order_node_t *node = order_head;
-    while (node != order_tail) {
-        if (strcmp(node->type, order_tail->type)) {
-            if (strstr(node->type, "BUY")) {
-                if (node->price >= order_tail->price) {
-                    if (deal_order(node, order_tail, node->price)) {
-                        break;
-                    }
-                }
-            } else {
-                if (node->price <= order_tail->price) {
-                    if (deal_order(node, order_tail, order_tail->price)) {
-                        break;
-                    }
-                }
+    if (buy_tail == NULL || sell_tail == NULL)
+        return;
+    order_node_t *node = buy_head;
+    while (node) {
+        if (node->price >= sell_tail->price) {
+            if (deal_order(node, sell_tail, sell_tail->price)) {
+                break;
+            }
+        }
+        node = node->next;
+    }
+    node = sell_head;
+    while (node) {
+        if (node->price <= buy_tail->price) {
+            if (deal_order(node, sell_tail, sell_tail->price)) {
+                break;
             }
         }
         node = node->next;
